+++ src/scripts/main.js
@@ -1,8 +1,8 @@
 import friendEvents from "./friends/eventListeners"
-import authorization from "./auth/eventListeners.js";
-import eventEvents from "./events/eventListeners.js";
-import articleEvents from "./articles/eventListeners.js";
-import taskEvents from "./tasks/eventListeners";
+import authorization from "./auth/eventListeners"
+import eventEvents from "./events/eventListeners"
+import articleEvents from "./articles/eventListeners"
+import taskEvents from "./tasks/eventListeners"
 import messageEvents from "./messages/eventListeners"
 
 /*
@@ -42,81 +42,10 @@ taskEvents.taskComplete();
 taskEvents.finishedTasks();
 taskEvents.standardTasks();
 
-//Friends List Object
-//****************************
-
-const friendListObject = {
-    fillFriendList: function (friendArray, mainUserNum) {
-        const friendListElement = document.querySelector("#friends-list")
-        friendArray.forEach(element => {
-            friendListElement.innerHTML +=
-                `
-            <div id = "friendCell-${element.id}" class = "friendCell"> 
-                <p>${element.userName}</p>
-                <button id = "delete-${element.id}" class = "deleteButton">Remove Friend</button>
-            </div>
-            `
-        });
-    },
-
-    addToFriendsList: function (event) {
-        const userIDAdded = event.target.id.split("-");
-        const addedFriend = {
-            userId: parseInt(userIDAdded[1]),
-            friendInitiate: 1
-        }
-
-        fetch("http://localhost:8088/friends", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify(addedFriend)
-        })
-            .then(data => {
-                return fetch(`http://localhost:8088/friends?userId=${addedFriend.userId}&friendInitiate=${addedFriend.friendInitiate}&_expand=user`)
-                    .then(newFriend => newFriend.json())
-                    .then(parsedFriend => {
-                        const friendListElement = document.querySelector("#friends-list")
-                        friendListElement.innerHTML +=
-                            `
-            <div id = "friendCell-${parsedFriend[0].id}" class = "friendCell"> 
-                <p>${parsedFriend[0].user.userName}</p>
-                <button id = "delete-${parsedFriend[0].id}" class = "deleteButton">Remove Friend</button>
-            </div>
-            `
-                        return parsedFriend
-                    })
-            })
-            .then(data => {
-                const buttonList = document.querySelectorAll(".deleteButton")
-                buttonList.forEach(element => {
-                    element.addEventListener("click", friendEvents.friendDelete)
-                });
-            })
-    }
-}
-
 //JM-Chatlog
 //**********************
 
 const chatObject = {
-    returnFriendArray: function (mainUserNum) { //Load function with the current user id
-        return fetch("http://localhost:8088/friends/?friendInitiate=1&_expand=user") //Fetch the friends of the user
-            .then(data => data.json())
-            .then(parsedData => {
-                let objectArray = []; //Array of users
-                parsedData.forEach(dataElement => {
-                    const friendObject = { //Collects the id and username of each user.
-                        id: dataElement.id,
-                        userNum: dataElement.user.id,
-                        userName: dataElement.user.userName
-                    }
-                    objectArray.push(friendObject);
-                });
-                return objectArray //Returns array
-            })
-    },
 
     returnMessagesArray: function (fetchedArray, mainUserNum) { //Returns the friend array and ID of current user
         //Populates the fetch string with multiple querys.
@@ -153,7 +82,7 @@ const chatObject = {
             })
             .then(parsedData => {
                 parsedData.forEach(element => {
-                    document.querySelector(`#message-${element.id} > .message-name`).addEventListener("click", friendListObject.addToFriendsList)
+                    document.querySelector(`#message-${element.id} > .message-name`).addEventListener("click", friendEvents.addToFriendsList)
                 });
                 const listOfEditButtons = document.querySelectorAll(".edit-button");
                 listOfEditButtons.forEach(element => {
@@ -197,11 +126,11 @@ document.querySelector("#submitChat").addEventListener("click", function () {
 
 var friendArray = []
 
-chatObject.returnFriendArray(1)
+friendEvents.returnFriendArray(1)
     .then(data => {
         friendArray = data;
         chatObject.returnMessagesArray(friendArray, 1);
-        friendListObject.fillFriendList(friendArray, 1);
+        friendEvents.fillFriendList(friendArray, 1);
         return friendArray;
     })
     .then(data => {
